+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/share_responses.hpp
@@ -244,6 +244,22 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     friend class PagedResponse<ListFileHandlesPagedResponse>;
   };
 
+  class ForceCloseAllFileHandlesPagedResponse
+      : public Azure::Core::PagedResponse<ForceCloseAllFileHandlesPagedResponse> {
+  public:
+    int32_t NumberOfHandlesClosed = 0;
+    int32_t NumberOfHandlesFailedToClose = 0;
+
+  private:
+    void OnNextPage(const Azure::Core::Context& context);
+
+    std::shared_ptr<ShareFileClient> m_shareFileClient;
+    ForceCloseAllFileHandlesOptions m_operationOptions;
+
+    friend class ShareFileClient;
+    friend class PagedResponse<ForceCloseAllFileHandlesPagedResponse>;
+  };
+
   class ListDirectoryHandlesPagedResponse
       : public Azure::Core::PagedResponse<ListDirectoryHandlesPagedResponse> {
   public:
@@ -257,6 +273,22 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
     friend class ShareDirectoryClient;
     friend class PagedResponse<ListDirectoryHandlesPagedResponse>;
+  };
+
+  class ForceCloseAllDirectoryHandlesPagedResponse
+      : public Azure::Core::PagedResponse<ForceCloseAllDirectoryHandlesPagedResponse> {
+  public:
+    int32_t NumberOfHandlesClosed = 0;
+    int32_t NumberOfHandlesFailedToClose = 0;
+
+  private:
+    void OnNextPage(const Azure::Core::Context& context);
+
+    std::shared_ptr<ShareDirectoryClient> m_shareDirectoryClient;
+    ForceCloseAllDirectoryHandlesOptions m_operationOptions;
+
+    friend class ShareDirectoryClient;
+    friend class PagedResponse<ForceCloseAllDirectoryHandlesPagedResponse>;
   };
 
 }}}} // namespace Azure::Storage::Files::Shares